:sparkles: Added session management
Introduced a new Session struct in the session package to handle session configuration. This includes cookie lifetime, persistence, name, domain, type and security settings. The InitSession function initializes a new session with these configurations. Default values are set for cookie lifetime if conversion fails and for session type if it's not specified as 'redis'.