:sparkles: Added JSON response utility
A new function `WriteJSON` has been added to the Celeritas package. This function is responsible for marshalling data into JSON format and writing it as a response. It also allows setting custom headers for the HTTP response. Error handling is included in case of any issues during marshalling or writing the response.