:sparkles: Added strcase library
The strcase library has been added to the project. This library provides functions for converting strings to various cases such as snake_case, camelCase, and kebab-case. It also includes a function for adding custom acronyms that need specific handling during conversion. The Travis CI configuration file for this library was also included in the commit.