🔒 Added security checks and code quality runner
Introduced two new GitHub workflows. The first one runs Gosec, a security scanner for Go projects, on every push and pull request to the main branch. The second workflow is triggered on pull requests and runs Trunk, a tool for checking code quality. It also cancels any in-progress run when a new one starts to save resources.