:sparkles: Added JSON read functionality
A new function `ReadJSON` has been added to the Celeritas package. This function takes in a ResponseWriter, Request and an interface, and returns an error. It reads JSON from the request body up to a maximum size of 1MB. If more than one JSON value is found in the body, it returns an error.